requirements.txt
easydict==1.9
Flask==2.1.0
gensim==4.2.0
jieba==0.42.1
numpy==1.19.5
python-dotenv==0.20.0
